move scanner graph_traversal to util/
