+++ depobs/worker/scoring.py
@@ -15,7 +15,7 @@ from depobs.database.models import (
     PackageReport,
     PackageVersion,
 )
-from depobs.scanner.graph_traversal import node_dep_ids_iter
+from depobs.util.graph_traversal import node_dep_ids_iter
 from depobs.util import graph_util
