+++ setup.cfg
@@ -55,7 +55,7 @@ disallow_untyped_calls = True
 disallow_incomplete_defs = True
 disallow_untyped_decorators = True
 
-[mypy-depobs.scanner.graph_traversal]
+[mypy-depobs.util.graph_traversal]
 check_untyped_defs = True
 disallow_untyped_defs = True
 disallow_untyped_calls = True
